Getting dates and floats working

+++ Assets/Plugins/iOS/CarnivalWrapper.h
@@ -27,7 +27,7 @@ extern "C" void _logEvent(const char *event);
 extern "C" void _setString(const char *string, const char *key);
 extern "C" void _setBool(bool boolValue, const char *key);
 extern "C" void _setDate(int64_t secondsSince1970, const char *key);
-extern "C" void _setFloat(double floatValue, const char *key);
+extern "C" void _setFloat(float floatValue, const char *key);
 extern "C" void _setInteger(int64_t intValue, const char *key);
 extern "C" void _removeAttribute(const char *key);
